backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pandas
openpyxl
python-multipart
passlib[bcrypt]
supabase
psycopg2-binary
dotenv